+++ src/server/physics/event/action-manager.ts
@@ -2,6 +2,7 @@ import { runTask } from 'cc-instanceinator/src/inst-util'
 import { assert } from '../../../misc/assert'
 import { prestart } from '../../../loading-stages'
 import { isPhysics } from '../is-physics-server'
+import { Client, runTaskInMapInst } from '../../../client/client'
 
 declare global {
     namespace ig {
@@ -9,6 +10,8 @@ declare global {
 
         interface ActorEntity {
             actionBoundToPlayer?: dummy.DummyPlayer
+
+            getClientFromBoundAction(this: this): Client | undefined
         }
     }
 }
@@ -22,6 +25,9 @@ prestart(() => {
     if (!PHYSICS) return
 
     ig.ActorEntity.inject({
+        getClientFromBoundAction() {
+            return this.actionBoundToPlayer?.getClient(true) ?? runTaskInMapInst(() => ig.ccmap?.clients[0])
+        },
         setAction(action, keepState, noStateReset) {
             if (action && isPhysics(multi.server) && !(this instanceof dummy.DummyPlayer)) {
                 const player = ig.actionNextTriggeredBy || ig.client?.dummy
@@ -32,7 +38,7 @@ prestart(() => {
         cancelAction(...args) {
             if (!this.currentAction || !this.actionBoundToPlayer) return this.parent(...args)
 
-            const client = this.actionBoundToPlayer.getClient(true)
+            const client = this.getClientFromBoundAction()
             if (!client) return this.parent(...args)
             return runTask(client.inst, () => this.parent(...args))
         },
@@ -42,7 +48,7 @@ prestart(() => {
         run(actor) {
             if (!actor.actionBoundToPlayer) return this.parent(actor)
 
-            const client = actor.actionBoundToPlayer.getClient(true)
+            const client = actor.getClientFromBoundAction()
             if (!client) return this.parent(actor)
             return runTask(client.inst, () => this.parent(actor))
         },