+++ src/types.ts
@@ -6,6 +6,7 @@ import type {} from 'nax-ccuilib/src/ui/pause-screen/pause-screen-api'
 import type {} from 'nax-ccuilib/src/ui/quick-menu/quick-menu-extension'
 import type {} from 'cc-variable-charge-time/src/plugin'
 import type {} from 'cc-krypek-lib/src/plugin'
+import type {} from 'cc-jetpack-widget/src/plugin'
 import 'ts-binarifier/src/type-aliases'
 
 declare global {