+++ src/state/entity.ts
@@ -41,6 +41,7 @@ import './entity/sc_FoodIconEntity'
 import './entity/ig_ENTITY_Door'
 import './entity/sc_BombEntity'
 import './entity/ig_ENTITY_BombPanel'
+import './entity/sc_FerroEntity'
 
 type EntityStateUnion = EntityStates[keyof EntityStates]