+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 ### Added
 
 - Add remote ig.ENTITY.Door animation
+- Add sc.BombEntity remote state
 
 ### Fixed