+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 - Add remote ig.ENTITY.Door animation
 - Add sc.BombEntity remote state
 - Add ig.ENTITY.BombPanel remote state
+- Add sc.FerroEntity remote state
 
 ### Fixed